Fix compiler warning on visual studio

+++ include/channeler/packet.h
@@ -192,7 +192,9 @@ struct private_header_layout
 struct footer_layout
 {
   // Footer offset is negative.
-  static constexpr ssize_t FOOT_OFFS_CHECKSUM = -sizeof(liberate::checksum::crc32_serialize); // CRC32
+  static constexpr ssize_t FOOT_OFFS_CHECKSUM = -static_cast<ssize_t>(
+      sizeof(liberate::checksum::crc32_serialize) // CRC32
+  );
 
   static constexpr size_t FOOT_SIZE = -1 * FOOT_OFFS_CHECKSUM;
 };
